requirements.txt
Bootstrap-Flask==1.5.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
filestack-python==3.4.0
Flask==2.0.1
Flask-WTF==0.14.3
fpdf==1.7.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.25.1
trafaret==2.0.2
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
